debug

+++ RecMST-Duplicate/src/RecMST.cu
@@ -217,7 +217,7 @@ void printColorArr(float* d_avg_color_r, float* d_avg_color_g, float* d_avg_colo
 	cudaMemcpy(h_avg_color_b, d_avg_color_b, sizeof(float) * n_elements, cudaMemcpyDeviceToHost);
 
 	for (int i = 0; i < n_elements; i++) {
-		printf("(%u, %u, %u) ", (unsigned int) h_avg_color_r[i], (unsigned int) h_avg_color_g[i], (unsigned int) h_avg_color_b[i]);
+		printf("(%f, %f, %f) ", h_avg_color_r[i], h_avg_color_g[i], h_avg_color_b[i]);
 	}
 
 	printf("\n");
